requirements.txt
requests==2.22.0
progressbar33==2.4
PyPDF2==1.26.0
